requirements.txt
streamlit 
yfinance 
pandas 
langchain-ollama
arize-phoenix 
openinference-instrumentation-langchain